fixed seed, fixed xts

+++ src/ciphers/seed/seed_expansion.h
@@ -2,10 +2,11 @@
 #define SEED_EXPANSION_H
 
 #include <stdint.h>
+#include <stddef.h>
 #include "seed_state.h"
 
 
-void seed_expandkey(uint64_t *user_key, uint64_t *key_enc, uint64_t *key_dec, size_t keylen);
+void seed_expandkey(uint32_t *user_key, uint32_t *key_enc, uint32_t *key_dec, size_t keylen);
 
 
 #endif
